fix: Fixed whole program
The changes were made in accordance with the change in the task conditions, namely parsing the input file, one thread parses and writes to the result.txt file, the other thread reads and sorts the received lines.
Now program works with two threads, which are using conditional variable and mutex.

Signed-off-by: Roman Veskera <roman.veskera@plvision.eu>